+++ src/scag/util/storage/Glossary.cpp
@@ -90,7 +90,7 @@ int Glossary::GetKeyByValue(int value, string& key) {
     smsc_log_debug(logger, "cancel GetKeyByValue - Glossary not opened");
     return GLOSSARY_CLOSED;
   }
-  if (value < glossVector.size()) {
+  if (size_t(value) < glossVector.size()) {
     key = glossVector[value];
     smsc_log_debug(logger, "GetKeyByValue. %d - %s", value, key.c_str());
     return SUCCESS;
